feat: Add API endpoint to get all users
This commit adds a new API endpoint `/api/getUsers` to retrieve all users from the database. It includes the necessary changes to the `index.js` file to register the new route. The `GetUsers` route file is also added to handle the request and return the users as a JSON response.

This feature enhances the API by providing a way to fetch all users, which can be useful for administrative purposes or displaying a list of users in the frontend.